First 1st Git push
The main aim of this project is to perform image classification based on data generated from satellites and provided to the security centers. Image classification can be used for several purposes, ranging from the detection of any security concerns in border areas to identifying any damaged or broken ships stuck up in the sea as well as locating any pirate ships in the sea that pose a threat to other voyagers.